move modified_cholesky to linalg
