+++ qiskit_nature/utils/__init__.py
@@ -25,16 +25,21 @@ Linear algebra utilities
 
    apply_matrix_to_slices
    givens_matrix
+   modified_cholesky
+
+Low rank decomposition utilities
+--------------------------------
    DoubleFactorizedHamiltonian
    low_rank_decomposition
 """
 
-from .linalg import apply_matrix_to_slices, givens_matrix
+from .linalg import apply_matrix_to_slices, givens_matrix, modified_cholesky
 from .low_rank import DoubleFactorizedHamiltonian, low_rank_decomposition
 
 __all__ = [
     "apply_matrix_to_slices",
     "givens_matrix",
+    "modified_cholesky",
     "DoubleFactorizedHamiltonian",
     "low_rank_decomposition",
 ]
